#301 로그, 리카나리 제거

+++ app/src/main/java/teamh/boostcamp/myapplication/view/setting/DeleteDiaryFileWorker.java
@@ -22,6 +22,9 @@ public class DeleteDiaryFileWorker extends Worker {
         File[] diaryFileList = diaryDirectory.listFiles();
 
         if (diaryDirectory.exists()) {
+            if(diaryFileList == null)
+                return Result.success();
+
             for (File diaryFile : diaryFileList) {
                 diaryFile.delete();
             }
